test(macos): one tests/common for the harness=false integration targets
The seven mac-gated targets in `crates/glass-macos/tests/` are
`harness = false`, so each has to do for itself what libtest would — format a
failure, exit non-zero, reap the fixture process. Each had grown its own copy:
`fail` in all seven, `swiftc_available` and `try_expect` in five, `build_fixture`
and `expect` in four, `with_stop_app` in three, the pixel helpers in two. Their
own comments conceded it — "identical to `capture.rs`'s/`input.rs`'s
`build_fixture`", "mirrors `tests/bundle_launch.rs`'s identically-named helper".

They now come from `tests/common/mod.rs`, which cargo does not treat as a target
because it is a subdirectory.

Two needed a parameter rather than a straight move. `build_fixture` takes the
fixture stem and a tag, because a11y builds a different `.swift` and each target
wants its own temp dir so a parallel run cannot have two building into one.
`with_stop_app` takes the label the two probes' copy had and bundle_launch's did
not; its three call sites now name the check they guard.

The largest was `run`, identical in capture, windows and input but for the pass
marker — verified mechanically, by normalising comments and the marker and
diffing. It is `run_fixture_test`, taking the checks as a closure, and it keeps
the cleanup ordering the copies documented: `stop_app` and the temp-dir removal
happen before any `process::exit`, which skips destructors.

601 lines out, 205 in.
